refactor(profile): change directory and name to combine with PBF
