+++ backend/src/test/java/com/brokendev/backend/controllers/UserProfileControllerTest.java
@@ -1,5 +1,6 @@
 package com.brokendev.backend.controllers;
 
+import com.brokendev.backend.profile.controller.UserProfileController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -7,9 +8,9 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.MockMvc;
 
-@WebMvcTest(UserController.class)
+@WebMvcTest(UserProfileController.class)
 @AutoConfigureMockMvc(addFilters = false)
-public class UserControllerTest {
+public class UserProfileControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
